Reorganise files and add individual scrolling for the two columns on the notes page
