+++ components/note/NoteCard.js
